% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsCor.R
\name{AbsCor}
\alias{AbsCor}
\title{Calculate absolute correlations among gene pairs}
\usage{
AbsCor(DataIn, method="pearson", diagNA=TRUE)
}
\arguments{
\item{DataIn}{input data, gene-by-sample matrix}

\item{method}{"pearson" or "spearman"; default is "pearson"}

\item{diagNA}{whether replace diagonal values to NA's}
}
\value{
Output is a gene-by-gene matrix; 
the i, j th entry shows the absolute correlation of the ith and jth gene.
}
\description{
Calculate absolute correlations among gene pairs
}
\examples{
AbsCor(matrix(rnorm(10),ncol=5))
}
\author{
Ning Leng
}
