% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_statistics.R
\name{filter_by_mean_count}
\alias{filter_by_mean_count}
\title{Filter genes by mean counts}
\usage{
filter_by_mean_count(gene_stat, low = 100, high = Inf)
}
\arguments{
\item{gene_stat}{A dataframe as returned by function
\code{\link[oscillation]{gene_statistics}}).}

\item{low}{Lower bound for the mean (use \code{-Inf} for an effective no
bound).}

\item{high}{Upper bound for the mean (use \code{Inf} for an effective no
bound).}
}
\value{
A dataframe of the same structure as \code{gene_stat} of filtered
  genes.
}
\description{
This functions takes a dataframe of gene expression statistics (see function
\code{\link[oscillation]{gene_statistics}}) and keeps only those genes whose
mean is within \code{low} and \code{high}, i.e., lie in the interval
[\code{low}, \code{high}].
}
