% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_sine_model.R
\name{find_optimal_phase_shift}
\alias{find_optimal_phase_shift}
\title{Optimal phase shift}
\usage{
find_optimal_phase_shift(x1, x2, psi_start = 0, method = "L-BFGS-B")
}
\arguments{
\item{x1}{Vector of gene expression values for gene 1.}

\item{x2}{Vector of gene expression values for gene 2.}

\item{psi_start}{The phase shift (psi) starting value for the optimisation
search.}

\item{method}{The name of the optimisation method to be used by
\code{\link[stats]{optim}}.}
}
\value{
A \code{\link[tibble]{tibble}} where rows are for each pair of genes,
  and five columns: gene 1, gene 2, \code{psi}, \code{se_epsilon} and
  \code{score}.
}
\description{
This function determines the optimal phase shift between two gene expression
profiles \code{x1} and \code{x2} according to the paired-sine model, see
\href{https://doi.org/10.1038/nmeth.3549}{Leng et al. 2015}, section
\emph{Oscope: paired-sine model}.
}
