% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_statistics.R
\name{gene_statistics}
\alias{gene_statistics}
\title{Gene expression statistics}
\usage{
gene_statistics(m, p_min = 0)
}
\arguments{
\item{m}{Matrix of gene expression values. Rows are genes and columns are samples.}

\item{p_min}{Minimum value for parameter p of the Negative Binomial
distribution to be assumed when estimating q.}
}
\description{
This function computes five statistics: mean, median, variance, \eqn{q}
(actually, \code{q1} and \code{q2}), and \eqn{\phi}. The statistics \eqn{q}
and \eqn{\phi} are parameter estimations related to the
\href{https://en.wikipedia.org/wiki/Negative_binomial_distribution}{Negative
Binomial} distribution, a commonly assumed model for the distribution of gene
(or isoform) expression counts.
}
