% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_statistics.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na_rm = TRUE, zero_propagate = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Whether to remove \code{NA} before calculation.}

\item{zero.propagate}{Whether zeros are allowed to propagate in the
calculation or whether they are treated as \code{NA}.}
}
\value{
The geometric mean.
}
\description{
This function computes the geometric mean.
}
