\name{Oscope-package}
\alias{Oscope-package}
\alias{Oscope}
\docType{package}
\title{
Oscope - A statistical pipeline for identifying oscillatory genes in unsynchronized single cell RNA-seq
}
\description{
Oscope is a statistical pipeline developed to identifying and recovering the base cycle profiles
of oscillating genes in an unsynchronized single cell RNA-seq experiment.
The Oscope pipeline includes three modules: a sine model module to search for
candidate oscillator pairs; a K-medoids clustering module to cluster candidate oscillators
into groups; and an extended nearest insertion module to recover the base cycle order
for each oscillator group.
}
\details{
\tabular{ll}{
Package: \tab Oscope\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2015-7-27\cr
License: \tab Artistic-2.0\cr
}
}
\author{
Ning Leng

Maintainer: Ning Leng <lengning1@gmail.com>
}
\references{
Leng et al. Oscope - A statistical pipeline for identifying oscillatory genes in unsynchronized single cell RNA-seq, accepted
}
\keyword{ package }
