\name{OscopeExampleData}
\alias{OscopeExampleData}
\docType{data}
\title{
	Simulated gene level data set with 600 genes and 30 cells.
}
\description{
	Simulated gene expression to evaluate Oscope. 600 genes and 30 cell are simulated.
	The expression mean of the genes are randomly simulated in the range of 10-10000.
}
\format{
	GeneExampleData is a matrix with 600 genes (rows) and 30 cells (columns). 
}
\examples{
data(OscopeExampleData)
str(OscopeExampleData)
}
\keyword{datasets}




