% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_sine_model.R
\name{paired_sine_analysis}
\alias{paired_sine_analysis}
\title{Perform paired-sine mode analysis}
\usage{
paired_sine_analysis(
  m,
  verbose = FALSE,
  sort_by_score = FALSE,
  parallel = FALSE,
  cores = parallel::detectCores() - 2L
)
}
\arguments{
\item{m}{Matrix of gene expression values. Rows are genes and columns are samples.}

\item{verbose}{Whether to be chatty.}

\item{sort_by_score}{Whether to descending sort the final dataframe by the
score column, i.e., by \eqn{-log10(sq_epsilon)}. If \code{FALSE}, the
returned dataframe will be sort by the original gene pair combinations.}

\item{parallel}{Whether to split the computation across the virtual cores.}

\item{cores}{The number of virtual cores (vCPUs) to use if \code{parallel}
equals \code{TRUE}.}
}
\value{
A dataframe with the optimal phase estimates. Each row corresponds to
  a gene pair combination. Columns:
\describe{
  \item{gene_1}{Name of gene 1.}
  \item{gene_2}{Name of gene 2.}
  \item{psi}{Optimal phase shift.}
  \item{sq_epsilon}{Squared epsilon (\eqn{\epsilon_{1,2}^2}).}
  \item{score}{Score defined as \eqn{-log10(\epsilon_{1,2}^2})). This metric
  can be used to sort the table of results, and select candidate oscillatory
  genes as those in the top gene pairs.}
}
}
\description{
Following the paired-sine model, as described in
\href{https://doi.org/10.1038/nmeth.3549}{Leng et al. 2015}, section
\emph{Oscope: paired-sine model}, this function estimates the optimal phase
shift for all gene pairs by minimising the squared epsilon error term, i.e.,
\eqn{\epsilon_{1,2}^2}.
}
