% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeR.R
\name{PipeR}
\alias{PipeR}
\title{Calculate residual of polynomial fit}
\usage{
PipeR(Data,Ndg=3,Method="Poly")
}
\arguments{
\item{Data}{gene-by-sample matrix or isoform-by-sample matrix.It should be rescaled to values bwteen
[-1,1].}

\item{Ndg}{degree of polynomial.}

\item{Method}{only polynomial fitting ("Poly") is available now.}
}
\value{
The function will fit polynomial curve to each row of the data. 
The output returns the MSE of each row (gene/isoform).
}
\description{
Calculate residual of polynomial fit
}
\examples{
aa <- sin(seq(0,1,.1))
bb <- sin(seq(0.5,1.5,.1))
cc <- sin(seq(0.9,1.9,.1))
res <- PipeR(rbind(aa,bb,cc))
}
\author{
Ning Leng
}
