% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_statistics.R
\name{pseudo_ref_sample}
\alias{pseudo_ref_sample}
\title{Calculate the pseudo-reference sample}
\usage{
pseudo_ref_sample(m)
}
\arguments{
\item{m}{Matrix of gene expression values. Rows are genes and columns are
samples.}
}
\value{
A vector of geometric means, one for each gene (can be regarded as
  pseudo-reference sample for each gene).
}
\description{
This function calculates the geometric mean across samples. This can be
regarded as a pseudo-reference sample. See
\href{http://genomebiology.com/2010/11/10/R106}{Anders & Huber (2010)}, page
3, denominator in equation 5.
}
