% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gsm_as_matrix}
\alias{read_gsm_as_matrix}
\title{Reads a GEO GSM (.csv or .csv.gz) file}
\usage{
read_gsm_as_matrix(filename)
}
\arguments{
\item{filename}{The path to the GSM file. It can read compressed (.gz) files.}
}
\value{
A matrix of gene expression values. Rows are genes/isoforms and
  columns are samples. Rownames are gene names.
}
\description{
This function reads a GEO GSM file with a matrix of expression data. Rows are
genes (or isoforms) and columns are samples. It expects a first row of column
names. The first column is then used to define the rownames of the matrix.
}
