% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_sine_model.R
\name{squared_epsilon}
\alias{squared_epsilon}
\title{Paired-sine model squared error term}
\usage{
squared_epsilon(psi, x1, x2)
}
\arguments{
\item{psi}{Phase shift between the gene expression of the two genes.}

\item{x1}{Vector of gene expression values for gene 1.}

\item{x2}{Vector of gene expression values for gene 2.}
}
\value{
Squared error term, i.e., \eqn{\epsilon_{1,2}^2}.
}
\description{
This function returns the squared error term as described in
\href{https://doi.org/10.1038/nmeth.3549}{Leng et al. 2015}, section
\emph{Oscope: paired-sine model}.
}
\details{
The definition of the error term is:
\deqn{\epsilon_{1,2}^2 = \sum_s ( x_{1,s}^2 + x_{2,s}^2 - 2 x_{1,s} x_{2,s} cos(\psi) - sin(\psi)^2)^2}
}
