% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_statistics.R
\name{median_norm_size_factors2}
\alias{median_norm_size_factors2}
\title{Median normalisation size factors (alternative)}
\usage{
median_norm_size_factors2(m, na_rm = FALSE)
}
\arguments{
\item{m}{Matrix of gene expression values. Rows are genes and columns are
samples.}

\item{na_rm}{A logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
A vector of size factors, one for each sample.
}
\description{
This function calculates the median normalisation size factors of a matrix of
gene expression counts following the code of the function
\code{\link[EBSeq]{MedianNorm}} when the parameter \code{alternative=TRUE}.
}
